Fixes another compiler warning.

+++ src/multiplayer.hpp
@@ -23,7 +23,7 @@ class game_data;
 namespace mp {
 
 // max. length of a player name
-const int max_login_size = 18;
+const size_t max_login_size = 18;
 
 /*
  * This is the main entry points of multiplayer mode.
